requirements.txt
Mastodon.py
PyYAML
